netlink-packet-route: VXLAN group/6, local/6 should use Vec<u8>
In order to be consistent with the whole project, VXLAN group, group6,
local, local6 properties should be typw Vec<u8> instead of u32/u128 as
they are IPv4/6 addresses.

Address is being reported as Vec<u8> IPv4 or IPv6, we could disucss if
this is correct or not but as this has been here for a long time we
should keep it to do not break use cases.

Signed-off-by: Fernando Fernandez Mancera <ffmancera@riseup.net>