vxlan: fix port-range attribute marshalling
Since we were not updating the `buffer` variable when writing to it, the "max"
part of the range was over-writing the "min", which caused the netlink message
to be considered invalid.

By adjusting the buffer to where we write, the netlink add message was
properly processed, and the link successfully created.

Unit tests were added to assert the expected behavior.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>