macsec: support MACsec interface
The patch is introducing MACsec interface support. In addition, it is
providing the InfoMacSec Nla. This Nla is gathering all the IFLA_MACSEC_
netlink attributes.

Please, notice that IFLA_MACSEC_PAD is omitted as it is used in SCI and
Cipher suite. It is not exposed in the nla_policy for MACsec.

Unit test added.

Signed-off-by: Fernando Fernandez Mancera <ffmancera@riseup.net>