Rename NetnsId to avoid surprises
There are 4 IFLA_* values related to network-namespaces.
The following 3 control the `dest_net` of the link:
 - `IFLA_NET_NS_PID`      by PID
 - `IFLA_NET_NS_FD`       by FD
 - `IFLA_TARGET_NETNSID`  by ID    (an alias for `IFLA_IF_NETNSID`)

And the last one controls the `link_net` of the link:
 - `IFLA_LINK_NETNSID`    by ID

Currently netlink-packet-route provides 3 enum variants with similar
names, which suggest they do the same thing with different input
 - `NetNsFd`    => Corresponds to IFLA_NET_NS_FD
 - `NetNsPid`   => Corresponds to IFLA_NET_NS_PID
 - `NetnsId`    => Corresponds to IFLA_LINK_NETNSID, not IFLA_TARGET_NETNSID

The similar naming might cause confusion, since it matches neither
the meaning nor the underlying CONSTANT_NAME used by the kernel.

Signed-off-by: Gilad Naaman <gilad@naaman.io>