fix: enable the latest rtnetlink on Android
The latest version of the rtnetlink crate cannot compile on Android.
The reason is that Android uses the AddressFamily defined in
address_family_fallback.rs, while the correct one should be that in
address_family_linux.rs. This leads to AddressFamily::Bridge used in
rtnetlink being undefined on Android.

This patch fixes this by adding necessary target_os="android" configs.
Since libc::AF_IB & libc::AF_MPLS are not defined on Android, we still
confine branches related to those enums to only linux and fuchsia.

Signed-off-by: chenzikai <chenzikai@google.com>