Add support for Seg6 encapsulation type
This commit provide some basic support for Seg6
(https://segment-routing.org/). Support for Inline
and Encap more are added. (Further support could be
enhanced in the future to support the remaining type,
but this is in my opinion a good starting point).
Additional tests are also included by comparing the
resulting messages with messaged captured in wireshark.

Note: two expect() are called in the code, but only
whenever it is logically impossible to have an exception
(checked bounds for IPv6 and parsing of a static IPv6
address)

Signed-off-by: Aperence <anthony.doeraene@hotmail.com>