attribute: fix decoding error on empty IFLA_VFINFO_LIST payload
Using netdevsim driver, it sends empty IFLA_VFINFO_LIST if not
configured. When this happens, the payload length is 0. As we are using
new_checked() function for NlaBuffer creation, we check that the payload
size can allocate the NLA size for that type.

The kernel shouldn't send an empty IFLA_VFINFO_LIST on netdevsim driver
when not configured but until that is fixed, this is a proper solution.

It fixes the following error:

```
thread 'main' panicked at links_dump.rs:62:52:
called `Result::unwrap()` on an `Err` value: DecodeError { inner: Failed to parse message with type 16

Caused by:
    Decode error occurred: invalid link message }
```

Signed-off-by: Fernando Fernandez Mancera <ffmancera@riseup.net>